* Change in behavior for search field: first result is now always selected; pressing enter/return on it will connect to that server. Would be nice to be able to press up/down in the search field to change server selection, but I think that would require subclassing the NSTextField * Server filter now works with connected servers as well as saved * Server filter now updates if you add or remove a server * Remove cruft in CRDSession; int ==> NSInteger replacements

+++ Source/CRDShared.h
@@ -100,8 +100,8 @@ NSNumber *CRDNumberForColorsText(NSString * colorsText);
 #pragma mark Constants
 
 // Constants
-extern const int CRDDefaultPort;
-extern const int CRDMouseEventLimit;
+extern const NSInteger CRDDefaultPort;
+extern const NSInteger CRDMouseEventLimit;
 extern const NSPoint CRDWindowCascadeStart;
 extern const float CRDWindowSnapSize;
 extern NSString * const CRDRowIndexPboardType;
